mpc-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
cvxpy==1.4.1
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1